flowgrind: add pareto distribution and unexpand whitespace
git-svn-id: svn+ssh://svn.umic-mesh.net/umic-mesh/projects/flowgrind/trunk@6524 3d4b872d-e00e-0410-a2ce-a5c2cda4747c

+++ src/fg_math.h
@@ -1,11 +1,11 @@
 /* initalization for random number generator */
-extern void init_math_functions (unsigned int seed);
+extern void init_math_functions (unsigned long seed);
 
 /* basic probability distributions */
-extern int 	dist_bernoulli	(const double p);
-extern double 	dist_pareto	(const double k, 	const double x_min); 
-extern double 	dist_weibull	(const double alpha, 	const double beta); 
-extern double 	dist_normal	(const double mu, 	const double sigma_square);
-extern double	dist_uniform	(const double minval, 	const double maxval);
-extern double	dist_exponential(const double mu);
-extern double 	dist_chisq 	(const double nu);
+extern int      dist_bernoulli  (const double p);
+extern double   dist_pareto     (const double k,        const double x_min);
+extern double   dist_weibull    (const double alpha,    const double beta);
+extern double   dist_normal     (const double mu,       const double sigma_square);
+extern double   dist_uniform    (const double minval,   const double maxval);
+extern double   dist_exponential(const double mu);
+extern double   dist_chisq      (const double nu);
